Module 2 Challenge
Included in the folder are 4 screenshots and the corresponding Excel sheet. The screenshots include the code used for the ticker counter and to calculate the open and close changes, as well as volume; the color code incorporated with the code from screenshot #1; the color code shown separately; and the code used to calculate the percent of increase and decrease as well as the volume for both "ticker" and "value". 
NOTE: the code did not save on multiple occasions, which is why there are screenshots of the individual codes. The spreadsheet shows that they all worked together, but wouldn't save in the Visual Basic tool. That issue has now been fixed.